Main Menu Prototype and Lane Minion AI Update
The main menu is finished with a functioning login server and database.
Gameplay wise, collisions are all around better handled, units are
easier to target with the mouse, and Lane Minions now keep a certain
distance from each other to make sure they don't stack so close that
they're untargetable.