End Game and Character Action UI
Added a "Victory" and "Defeat" screen that pops up at the end of the
game. Game characters now keep track of the actions being performed by
skills and have replicated timers to notify clients of actions currently
happening. The player HUD now has a widget that monitors the player
character's action and shows itself when it has one.