Bugfixes, Realm Turret AI Target Selection Update, and UI
Fixed a bug where effects weren't being removed when a unit dies. Fixed
a bug where the game didn't have a valid player state to use for a kill
notify. Turret target selection has been updated. Life hits are tracked
for a period of time after one has been added to a player character.
This means when a player dies, the last player controller to inflict
damage to the character is going to get the credit for the kill if they
were the last one to damage them in the last 7.5 seconds.