Fog of War Updates and Bugfixes
Fog of War now takes line of sight into account when calculating
visibility. The system now traces from the calculating character to the
possible in sight target. If it can trace without hitting anything, it
is counted as visible.

Fixed a bug where clients could send nullptrs as current targets to the
server and be kicked for failing validation. Fixed a bug where the game
would crash if an object was selected by the player that wasn't a valid
actor.