Bugfixes, Objective Deaths, and Objective Levelling System
Fixed a bug where the replication system crashed trying to replicate a
character's current target when it was being garbage collected. Current
targets now have to pass the IsValid() test to be accepted as a target.

Objectives now have functions to call in the game mode when they die
that way they can award team members, give stats, etc.

Objectives and minions now level with the average level of the players
on the team they are on. They also gain a level every certain period of
time.