Blueprint Damage Hooks for the Mods System, Combat State, and Effects Update
Added Blueprint hooks to call in the mods for when the player takes
damage.

Using skills, taking damage, and giving damage now all make a game
character enter their combat state. Their blueprints are notified
whenever this state starts or ends.

Effects can now be transferred on the death of their owner if the flag
is set. This is useful for transferring buffs from one player to another
on a kill.