Negate Character Damage, Forest Minion AI, and Mods
Added a flag to negate damage for game characters. This is useful for
mods that help mitigate incoming damage for characters.

Fixed a bug where select unit under mouse was crashing since it was
trying to access nullptrs.

Forest minions now communicate with their camp spawner and tell it to
set the respawn timer when the camp has been killed.

Mods now have a working recipe system. Mods that are included in the
recipes of other mods now count toward that mod.