Bugfixes, Mod Recipe UI, Respawn Times and Tooltips
Fixed a bug where auto attacks weren't being brought off cooldown if the
attacker died while the attack was on cooldown.

The in-game store now has a recipe component for displaying the needed
mods and credits to purchase the selected mod.

Respawn times now depend on more than just time. It depends on the
character's level and match length now. The player's camera now plays a
black and white effect when they are waiting to respawn.

Mods in the player HUD now display tooltips for information about the
mod the player is currently hovering over with their mouse.