Bugfixes and Shouta Support
Fixed a bug where a damage event was being sent parameters that were
invalid and null.

Added a shield overlay to player character health bars.

Made game characters have a targetable boolean and now characters can't
be targeted if that boolean is set to false.

Using skills now correctly replicate the proper trace hit information to
the server.

Fixed a bug in the shield system where removing shields wouldn't remove
their amount from the total shield amount.

Characters now can toggle when specific characters in a stored array can
be the only ones in the game to be able to damage them.