Overhead Widgets and Hover Targets
There is now a special class for UMG widgets to be spawned overhead of
game characters. This is to replace having to draw player info in code
with designing widgets in the editor.

Hovering over game targets in-game with your mouse now lights them up
cyan or red, depending if they're friendly units or not.