Minimap Prototype and Another Fog of War Update
Initial prototype of the minimap system in C++ (doesn't work to except
to show the owning player's position). Huge Fog of War update where it
now calculates visibilities on a separate thread. Boosts FPS a little
but still lags on the standard 3 lane map in-editor. Need to test
standalone client.